refactor: lift connection tab logic to community view
